% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\docType{methods}
\name{Snowflake}
\alias{Snowflake}
\alias{dbConnect,SnowflakeDriver-method}
\title{Snowflake DBI wrapper}
\usage{
Snowflake()

\S4method{dbConnect}{SnowflakeDriver}(drv, account_name,
  region_id = "us-west-2", user, password, db = "nodb",
  schema = "noschema", warehouse = "nowarehouse", ...)
}
\arguments{
\item{drv}{An object created by \code{Snowflake()}}

\item{account_name}{Snowflake account name, which is the subdomain part of your Snowflake warehouse}

\item{region_id}{id of the region assigned to your Snowflake account}

\item{user}{Snowflake username}

\item{password}{Snowflake password}

\item{db}{default database to use}

\item{schema}{default schema to use}

\item{warehouse}{default warehouse / compute cluster to use}

\item{...}{other parameters passed to the JDBC driver}
}
\description{
Snowflake DBI wrapper

Connect to Snowflake
}
\examples{
\dontrun{
require(DBI)
con <- dbConnect(AWR.Snowflake::Snowflake(),
  account_name = 'foobar', user = 'username', password = '***', db = 'SNOWFLAKE_SAMPLE_DATA')
dbListTables(con)
dbGetQuery(con, 'SELECT COUNT(*) from TPCDS_SF100TCL.call_center')
dbGetQuery(con, 'SELECT * FROM tpch_sf1.lineitem limit 5')
dbDisconnect(con)
}
}
\seealso{
\href{https://docs.snowflake.net/manuals/user-guide/jdbc-configure.html}{Snowflake JDBC Manual} for more connections options.
}
