% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowflake.R
\docType{class}
\name{SnowflakeDriver-class}
\alias{SnowflakeDriver-class}
\alias{SnowflakeDriver}
\title{Snowflake driver class.}
\usage{
\S4method{initialize}{SnowflakeDriver}(.Object, ...)
}
\description{
Snowflake driver class.

Constructor of SnowflakeDriver
}
\keyword{internal}
